Module nav was added
