--- src/app/module-viewer/riverside-step-template/templates/radio-button/Item.ts
@@ -1,6 +1,0 @@
-export class Item {
-    id?: number;
-    title: string;
-    description: string;
-    image: string;
-}
