+++ src/app/module-viewer/riverside-step-template/templates/radio-button/index.ts
@@ -1,5 +1,11 @@
 import { TemplateContentDataBase } from '../template.interface';
 
+export class Item {
+    id?: number;
+    title: string;
+    description: string;
+    image: string;
+}
 
 export interface RadiobuttonTemplateData extends TemplateContentDataBase {
     template_params_json: {
